refactor: modularize utility logic, update UI assets, and improve code structure
