--- src/index.css
@@ -1,9 +1,0 @@
-/* Theme  */
-
-:root {
-  --color-dark-grey: #343c40;
-  --color-clear-grey: #d8d9d7;
-  --color-grey: #a6a6a6;
-  --color-red: #fd5b36;
-  --color-black: #0d0d0d;
-}
