+++ src/utils/constants.ts
@@ -1,0 +1,6 @@
+export const HP_CLASS = "hp-disabled";
+export const HP_SPINNER_ID = "hp-loading-spinner";
+export const HP_HIDE_CLASS = "hp-hide-completely";
+export const HP_PLACEHOLDER_CLASS = "hp-hidden-placeholder";
+export const HP_REHIDE_BTN_CLASS = "hp-rehide-btn";
+export const HP_STYLE_ID = "hp-style";
